// src/types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Data {
  /**
   * Unique token ticker for indexing
   */
  ticker: string;
  shillers: string[];
  /**
   * Array of message objects
   */
  messages: {
    /**
     * Must be a string and is required
     */
    url: string;
    /**
     * UTC date of the message. Required.
     */
    date: number;
    /**
     * Must be a string and is required
     */
    content: string;
    /**
     * Must be a string and is required
     */
    author: string;
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}